Update how mirador item form indexes are handled to not use element length
If you remove the first item, the length is now two, however, w/o renaming
all form elements the last item index is also two, so we end up getting a clashing form name.

This commit changes the index handling to look at the indexes present in the form names, sort them, and ensure that we're incrementing.